+++ duoStrength.js
@@ -4395,8 +4395,19 @@ async function init()
 				// there is a topBarDiv so we can continue to process the page to workout what to do
 
 				// set username via the href of a link to the profile
-				let profileLinkHrefParts = document.querySelector(`[href^="/profile/"]`).href.split("/");
-				username = profileLinkHrefParts[profileLinkHrefParts.length - 1];
+				const profileLinkElement = document.querySelector(`[href^="/profile/"]`);
+				if (profileLinkElement === null)
+				{
+					// no element with the username in has been found
+					console.error("Username could not be pulled from the page:\nNo link to the user's profile found");
+					return false;
+				}
+				else
+				{
+					const profileLinkHrefParts = document.querySelector(`[href^="/profile/"]`).href.split("/");
+					username = profileLinkHrefParts[profileLinkHrefParts.length - 1];
+				}
+
 
 				// topBar Div is the direct container holding the navigation butons, has class _3F_8q
 				// old method topBarDiv = dataReactRoot.childNodes[2].childNodes[1].childNodes[2].childNodes[0];